Fix empty compile_commands.json
This commit re-enables the keep_last_cmd member variable in the TaskBase class
(which was disabled by default in Waf 2.0.9) when clang_compilation_database.py
is used. The clang_compilation_database.py file was correctly updated in the
relevant Waf commit introducing the change [1], but is missing in ns-3's bundled
version of Waf.

[1] https://gitlab.com/ita1024/waf/commit/7778868718333932708a3915825f5c3f4ac53341