lte: Do not generate CQI reports & UE measurements if the UE is not in synch with any cell
We do not generate the CQI report when the UE is not synchronized to any cell. Also, the RLF is detected after the DL CTRL is received by the UE,therefore, we do not need to generate the CQI reports and the UE measurements for a CTRL for which the RLF has been detected.